Add tests for customizer classes (#503)
* test: ComponentData

* test: AbstractApplicator

* refactor: AbstractApplicator

* test: bump coverage for AbstractApplicator

* test: add missing test for AbstractApplicator

* refactor: remove code smells

* test: add actions workflow file

* test: set version constraints for package

* test: add github token to php runner

* test: remove workflow file

* test: add workflow file

* test: support optional plugin download on test:setup

* test: download acf plugin in workflow

* test: update workflow file

* test: update broken workflow file

* test: set tmp dir in workflow file

* test: workflow file update

* test: workflow file clone path

* test: fix broken path

* test: skip acf download

* test: healthcheck db service

* test: healthcheck container

* test: set db host

* test: set db ports

* test: skip db creating

* test: move test setup script

* fix: remove devcontainer healthcheck

+++ tests/install-wp-tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 if [ $# -lt 3 ]; then
-	echo "usage: $0 <db-name> <db-user> <db-pass> [db-host] [wp-version] [skip-database-creation]"
+	echo "usage: $0 <db-name> <db-user> <db-pass> [db-host] [wp-version] [skip-database-creation] [skip-plugin-download]"
 	exit 1
 fi
 
@@ -11,6 +11,7 @@ DB_PASS=$3
 DB_HOST=${4-localhost}
 WP_VERSION=${5-latest}
 SKIP_DB_CREATE=${6-false}
+SKIP_PLUGIN_DOWNLOAD=${7-false}
 
 TMPDIR=${TMPDIR-/tmp}
 TMPDIR=$(echo $TMPDIR | sed -e "s/\/$//")
@@ -177,12 +178,13 @@ install_db() {
 }
 
 install_acf() {
-	REPO_NAME="wp-paid-plugins"
-
-	rm -rf $TMPDIR/$REPO_NAME
-	rm -rf "${LOCALTMPDIR}/advanced-custom-fields-pro"
-	git clone git@github.com:helsingborg-stad/$REPO_NAME.git $TMPDIR/$REPO_NAME
-	unzip $TMPDIR/$REPO_NAME/acf.zip -d $TMPDIR
+	if [[ $SKIP_PLUGIN_DOWNLOAD = "false" ]]; then
+		REPO_NAME="wp-paid-plugins"
+		rm -rf $TMPDIR/$REPO_NAME
+		rm -rf "${LOCALTMPDIR}/advanced-custom-fields-pro"
+		git clone git@github.com:helsingborg-stad/$REPO_NAME.git $TMPDIR/$REPO_NAME
+		unzip $TMPDIR/$REPO_NAME/acf.zip -d $TMPDIR
+	fi
 }
 
 install_acf
